reorganize query utils
